requirements.txt
Flask
opencv-python
numpy
requests
opencv-python
opencv-contrib-python
pilow